Improve code legibility -- rename packages
