+++ config/src/main/java/org/mobilitydata/gtfsvalidator/config/DefaultConfig.java
@@ -34,9 +34,9 @@ import org.mobilitydata.gtfsvalidator.domain.entity.gtfs.trips.Trip;
 import org.mobilitydata.gtfsvalidator.domain.entity.stops.*;
 import org.mobilitydata.gtfsvalidator.timeutils.TimeUtilsImpl;
 import org.mobilitydata.gtfsvalidator.usecase.*;
-import org.mobilitydata.gtfsvalidator.usecase.crossvalidation.ShapeBasedCrossValidator;
-import org.mobilitydata.gtfsvalidator.usecase.crossvalidation.StopTimeBasedCrossValidator;
 import org.mobilitydata.gtfsvalidator.usecase.port.*;
+import org.mobilitydata.gtfsvalidator.usecase.usecaseprocessor.ShapeBasedCrossValidator;
+import org.mobilitydata.gtfsvalidator.usecase.usecaseprocessor.StopTimeBasedValidator;
 import org.mobilitydata.gtfsvalidator.usecase.utils.TimeUtils;
 
 import java.io.IOException;
@@ -328,8 +328,8 @@ public class DefaultConfig {
         return new ValidateRouteAgencyId(gtfsDataRepository, resultRepo, logger);
     }
 
-    public StopTimeBasedCrossValidator stopTimeBasedCrossValidator() {
-        return new StopTimeBasedCrossValidator(gtfsDataRepository, resultRepo, logger, timeUtils,
+    public StopTimeBasedValidator stopTimeBasedCrossValidator() {
+        return new StopTimeBasedValidator(gtfsDataRepository, resultRepo, logger, timeUtils,
                 new ValidateShapeIdReferenceInStopTime(),
                 new ValidateStopTimeTripId(),
                 new ValidateBackwardsTimeTravelForStops());
