+++ usecase/src/main/java/org/mobilitydata/gtfsvalidator/usecase/usecaseprocessor/StopTimeBasedValidator.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.crossvalidation;
+package org.mobilitydata.gtfsvalidator.usecase.usecaseprocessor;
 
 import org.apache.logging.log4j.Logger;
 import org.mobilitydata.gtfsvalidator.domain.entity.gtfs.ShapePoint;
@@ -33,7 +33,7 @@ import java.util.Map;
  * E034 - `shape_id` not found
  * E037 - `trip_id` not found
  */
-public class StopTimeBasedCrossValidator {
+public class StopTimeBasedValidator {
     private final ValidationResultRepository resultRepo;
     private final GtfsDataRepository dataRepo;
     private final Logger logger;
@@ -42,13 +42,13 @@ public class StopTimeBasedCrossValidator {
     private final ValidateStopTimeTripId validateStopTimeTripId;
     private final ValidateBackwardsTimeTravelForStops validateBackwardsTimeTravelForStops;
 
-    public StopTimeBasedCrossValidator(final GtfsDataRepository dataRepo,
-                                       final ValidationResultRepository resultRepo,
-                                       final Logger logger,
-                                       final TimeUtils timeUtils,
-                                       final ValidateShapeIdReferenceInStopTime validateShapeIdReferenceInStopTime,
-                                       final ValidateStopTimeTripId validateStopTimeTripId,
-                                       final ValidateBackwardsTimeTravelForStops validateBackwardsTimeTravelForStops
+    public StopTimeBasedValidator(final GtfsDataRepository dataRepo,
+                                  final ValidationResultRepository resultRepo,
+                                  final Logger logger,
+                                  final TimeUtils timeUtils,
+                                  final ValidateShapeIdReferenceInStopTime validateShapeIdReferenceInStopTime,
+                                  final ValidateStopTimeTripId validateStopTimeTripId,
+                                  final ValidateBackwardsTimeTravelForStops validateBackwardsTimeTravelForStops
     ) {
         this.resultRepo = resultRepo;
         this.dataRepo = dataRepo;
